refactor: mover error codes to types
