+++ packages/velocity-exchange-verifiers/src/impl/errors/error-builder.ts
@@ -4,8 +4,7 @@
  * Copyright 2022 Velocity Career Labs inc.
  * SPDX-License-Identifier: Apache-2.0
  */
-import { VerificationError } from 'api/types';
-import { ErrorCode } from './error-code';
+import { ErrorCode, VerificationError } from 'api/types';
 
 /**
  * Creates a {@link VerificationError} with a given error code, message, and location path.
