+++ packages/velocity-exchange-verifiers/src/impl/verifiers/pure-verifiers/credential-status-verifier.ts
@@ -5,8 +5,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { buildError, ERROR_CODES } from 'impl/errors';
-import { W3CCredentialJwtV1, Verifier } from 'api/types';
+import { buildError } from 'impl/errors';
+import {W3CCredentialJwtV1, Verifier, ERROR_CODES} from 'api/types';
 
 /**
  * Verifies that the `vc.credentialStatus` field exists in the Credential JWT payload.
