+++ packages/velocity-exchange-verifiers/src/impl/verifiers/pure-verifiers/iss-claim-matches-either-metadata-or-credential-issuer-verifier.ts
@@ -4,8 +4,8 @@
  * Copyright 2022 Velocity Career Labs inc.
  * SPDX-License-Identifier: Apache-2.0
  */
-import { W3CCredentialJwtV1, Verifier, VerificationContext } from 'api/types';
-import { buildError, ERROR_CODES } from 'impl/errors';
+import {W3CCredentialJwtV1, Verifier, VerificationContext, ERROR_CODES} from 'api/types';
+import { buildError } from 'impl/errors';
 
 /**
  * Verifies that the Credential JWT's `iss` claim matches the expected issuer metadata.
