+++ packages/velocity-exchange-verifiers/src/impl/verifiers/pure-verifiers/kid-claim-is-velocity-v2-verifier.ts
@@ -5,8 +5,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { W3CCredentialJwtV1, Verifier } from 'api/types';
-import { buildError, ERROR_CODES } from 'impl/errors';
+import {W3CCredentialJwtV1, Verifier, ERROR_CODES} from 'api/types';
+import { buildError } from 'impl/errors';
 
 /**
  * Verifies that the Credential JWT's `kid` (Key ID) claim starts with the required Velocity DID prefix.
